+++ CHANGES.md
@@ -5,6 +5,7 @@
  * Remove all code when JENT_CONF_DISABLE_LOOP_SHUFFLE is unset. This code is already discouraged for a long time. Now it is taken out for good.
  * If cache size cannot be detected from base system (e.g. virtualization), use the requested memory size.
  * Change the stuck test to always calculate the absolute values of the 2nd and 3rd discrete derivation of time.
+* Replace SHA3-256 output generation with XDRBG-256 (common case) or SHA3-512-based XDRBG-like construction
 
 3.6.3
  * Correct time stamp processing on AIX